feat (api): improve loadUsedVariables performance
